+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ☮️ Peace in the world, or the world in pieces. 🕊️
+### Fixed
+- Various bugs in Fear table function.
 
 ## [2.6.1] - 2024-05-12
 ### Fixed